OMAR-DuplicateEncoder (#168)
* OMAR-DuplicateEncoder

The goal of this exercise is to convert a string to a new string where each character in the new string is '(' if that character appears only once in the original string, or ')' if that character appears more than once in the original string. Ignore capitalization when determining if a character is a duplicate.

Examples:

"din" => "((("

"recede" => "()()()"

"Success" => ")())())"

"(( @" => "))(("

Notes:

Assertion messages may be unclear about what they display in some languages. If you read "...It Should encode XXX", the "XXX" is actually the expected result, not the input! (these languages are locked so that's not possible to correct it).

* CESAR ENCRYPTION [FINAL KATA]

- Enconde ready.
- Decode ready.
- improve code and the Vigenere Encryption coming soon.

* Fix sorry this is my mistake

Fix sorry this is my mistake